moved the editor app class to a different file, added a basic test to see if the initialize function works
